test(mcp): split Plexus route coverage
